feat: accomplish the functions of store rule

+++ src/main/java/team/delete/scheduling_system/entity/Store.java
@@ -29,5 +29,5 @@ public class Store {
     /**
      * 占地面积
      */
-    private float size;
+    private Double size;
 }
